feat: Use a pool to keep extension contexts alive across requests.
This is neccessary since initializing Graal Polyglot Contexts is kind of
expensive, *especially* for Python, where we're looking at multi-second
startup times for the contexts. By using a pool we can avoid this most
of the time and serve clients with reduced latency.

This complicates the client API a bit: Extensions must now implement a
`cleanup` hook that cleans up any state accumulated while processing a
request. This hook is mandatory, just like the `info` hook, to prevent
clients from forgetting about this.