feat: Run auth and resolve hooks concurrently to reduce latency
When multiple auth or resolving extensions are active, we run them in
parallel to reduce the latency for the clients.

In the case of auth extensions, we run until the first hook returns a
negative value, in which case all remaining extension calls are
cancelled and we return.

In the case of resolving extensions, we run until first hook returns a
non-null/undefined result, in which case all remaining extension calls
are cancelled as well, and we return the resolving result.

Calling Graal Contexts from multiple threads (neccessary since we create
the Context on the servlet request thread and execute the auth/resolving
hooks in a thread from a shared pool) is possible, but requires using
the `.enter()/exit()` APIs and securing the access to the context with
a lock. To reduce the pain of that, we define two new APIs on
`RuntimeContext`:

- `public ContextLease enter()`: Returns an `AutoCloseable` lease on the
  context that provides access to the polyglot `Value` with the
  extension hook object
- `public <R> R run(Function<ContextLease, R> callable)`: Convenience
  API for running lambdas with a lease on the context

Consequently, `RecordContext` was promoted from a model record to a
"proper" class, that prevents direct access to the extension `Value`.