refactor: Split RuntimeContext into api+impl to get rid of CGLIB proxies
Spring can avoid creating hacky CGLIB proxies for certain beans if the
Bean type implements at least one interface, so it can create a
proxy implementation without relying on reflection magic, which tends to
not play well together with the Java Module System.

This change implements this for the `RuntimeContext` type, which
requires a proxy due to being a request-scoped bean. We change
`RuntimeContext` to be an interface, with the inner class
`RuntimeContextImpl` as its sole implementation.