feat: Validate configured extensions at startup
Delays readiness of application to check all enabled extensions for
compliance with the IIIF Image API specifications.

Runs the full validation test suite against each extension in isolation,
terminating startup with details if any of the tests fail.

Required adding a new API to `ExtensionRegistry` to temporarily disable
all extensions except for the one to be validates, implemented using an
`AutoCloseable` that restored the full set of extensions when closed
(i.e. when the validation run has terminated for the isolated extension)