feat: Allow overriding spring-settings from wolpi config
First change in a series where we try to make the end-user experience as
spring-free as possible, i.e. Spring/Spring Boot is an implementation
detail and users should not have to look into the Spring documentation
ever (except if they have very advanced use cases).

To do this, we introduce a new `ApplicationListener` class that hooks
into an early phase of the Spring startup process and sets various
spring properties based on values in the `wolpi` section.

This change implements this for the new `logging` and `http` sections,
that allow setting the log format (`text` or `json`) and the host, port
and request thread pool settings, as well as the base URI (for cases
where forwarding headers are not present on proxied requests, currently
not implemented in the controller).

Also adds a cute ASCII art banner.