feat: Add 'validate' subcommand
This subcommand will validate a local extension (provided as a path to a
file or pacakge directory) against the IIIF Image API validation test
suite. It includes a "watch" mode (enabled with `-w`/`--watch`) that
monitors the extension code for changes, reloads the extension and then
runs the test suite.

The reloading required a few changes to the `ExtensionRegistry` and
`*Installer` classes:

- `ExtensionRegistry#loadExtension` directly registers the extension
  instead of just returning a `LoadedExtension`
- Before registering an extension, a previously loaded version of the
  extension is unloaded from the registry to allow for hot reloading.
  Identity is determined by the extension's `name`.
- For both Python and JavaScript, local packages are now installed in
  "editable" mode, where changes to the code are directly reflected
  in the install-directory.

Some drive-by changes were also made:
- Some more tuning of the text logging with nicer emojis, a nicer
  startup info and less verbose logging of spring internals.
- Conversion into `URI` types from Graal guests has been made more
  robust
- Fixes to startup validation that were not noticed before