fix: Mapping of YAML strings to VNamedEnum values.
Previously we used Guava for this, but it turned out that its
implementation of class enumeration was incomplete and missed a lot of
Vips classes. We now use the `classgraph` library, which has a much more
comprehensive set of heuristics that also handles Spring Boot-specific
stuff, and now we are able to discover all of the available VNamedEnum
implementations we need for proper mapping of enum configuration
options.