feat: Support live-reloading of extensions.
This was imlemented half-way in the previous commit, but had some
problems that made it impractical (→ we were not busting Graal's
internal module cache, so JS changes were not picked up). This change
set fixes this and generalizes the reload feature to apply to loaded
extensions in general, i.e. users can set `liveRelad: true` in their
configuration for an extension and it will be reloaded at runtime
whenever the underlying code changes, making for a very short feedback
cycle when developing extensions.

Modifications that were made towards that goal:
- Check for environment support for editable installs in GraalJS and
  GraalPy (npm >= 10 for GraalJS, GraalPy standalone for GraalPy) when
  live reloading is enabled for an extensions, disable feature if
  prerequisites are not met
- Implement cache busting for ESM modules in GraalJS by importing packages
  with a `?ts=<timestamp>` query parameter, that gets ignored when
  accessing the filesystem (in `ESMFileSystem`), but is used by Graal as
  a cache key in the internal ESM module cache.
- `LoadedExtension` records have a new `lastModified` field that becomes
  part of their identity. This ensures that loaded extensions receive a
  new sub-pool in the pool for `RuntimeContext`s and no stale contexts
  are handed out.
- Importing the extension hooks from JS/Python packages got a bit more
  complicated to allow cache bustin