feat: Augment info.json responses from extensiosn
Difference compared to concept: Single hook `augmentInfoJson` with a parameter
for the IIIF versioninstead of two hooks, `augment` is a better fit for
the purpose than `postProcess` (→ extensions are mainly supposed to extend the
info.json data structure, not modify it arbitrarily if not strictly
neccessary).

Required adding a new rule to the architecture tests so that
`Controller` can access `Extension Layer` to invoke the hook.