fix: Fix passing data between different guest contexts.
For the info.json augmentation hook, we need to pass a
`dict/object/HashMap` not just between host and guest, but also between
guests (i.e. one extension takes the output of another as an input).
This revelead a mistake in the implementation, were we wrapped every
`Map` from the host in a `ProxyObject`, but that makes working with the
data structure in Python kind of painful. So now we use the most
idiomatic data type for each environment when dealing with JSON-ish data:

- Java: `Map<String, Object>`
- Python: `dict[str, Any]`/`typing.Mapping[str, Any]`
- JavaScript: `object`

`PolyglotHelpers#toGuest` was updated to perform that mapping.