feat: Add support for debugging extensions via ADB protocol
The Graal runtime can expose a debugging interface via the ADB protocol
that is supported by VSCode (native) and IntelliJ (via a plugin[1]).

Since debugging must be enabled for the complete engine, we had to make
the `GraalContextSupplier` a Spring Bean (to get access to the
configuration via DI), which is why the line count in this change is
rather high, especially in the tests.

Updated developer documentation as well.

Implemented a drive-by fix for guest strings always being converted to
`URI` instances due to a bad predicate, we now check if a string is a
HTTP/HTTPS URL before converting to `URI`.

[1] https://plugins.jetbrains.com/plugin/23257-lsp4ij