feat: More idiomatic access to extension guest context fields.
Previously, extensions had to invoke getter methods to access the
individual fields on the extension guest context object, which is
unidiomatic in both Python and JavaScript, where direct access to object
members is the established pattern.

To address this, we introduce a new class `RecordProxy` that wraps a
Java `Record` instance and offers idiomatic access through the Graal
`ProxyObject` API. This is hooked into the established
`PolyglotHelpers#toGuest` API.

The sample and test extensions were updated to use direct field access
instead of getter invocations.