feat: Python extensions should use an import to access the guest context
Builtins in Python are usually reserved for proper language builtins,
and unlike JavaScript it's not an expected pattern to provide custom
objects in this scope.

To improve the UX for Python extensions, we now provide access to the
guest context via a `wolpi` module that can be imported. This should
also make it easier to provide type hints later on.