refactor: Move guest context creation into its own factory bean
`ExtensionRegistry` already is pretty much a "god" object with all of
its resonsibilities, and a lot of DI-ed objects were just there to
create guest contexts. Moving this responsibility out of the bean into
its own factory doesn't solve the complexity in the registry, but it's
a little less state that we are coupled to.