fix: When calling extension hooks, make sure all arguments are converted to guest values
Previously we would simply provide the host values, including `Record`
objects. This led to an inconsistency in the API, where guests would
access fields on an object both as a simple member access (when
accessing the guest context, e.g.), and in other cases had to invoke
getter methods. This fix should remedy that at least for `Record`
parameters.