feat: Add a custom global error handler for more error context
Instead of a generic error message and a stack trace, we print:
- the request method and path for the request
- the remote address for the request
- the request headers for the request.

This should allow more precise debugging of crashes.
Logging the remote address and headers can be disabled for environments
where thesevalues are considered PII.

Also, we now heavily filter stack traces in the text log to hide noisy
frames from the JVM, Spring, Tomcat and Graal to make them easier to
parse.