Allow pre-installation of extensions into container.
To further reduce startup times for the Wolpi container, users can
customize the image by extending the base image, adding their extensions
and a config that references these extensions into the customized
container and then run the `install-extensions` subcommand.

This will then install the extensions into the container and validate
them. The resulting image should then be able to start up in
milliseconds, since no extensions need to be installed an no
verifications need to performed.

This should also serve as a good CI step for Wolpi deployments, if users
use this approach to pre-bake their extensions in the container, they
can be sure that there are no surprises when the container is started,
since no external sources need to be consulted (i.e. to fetch the
extensions).